+++ firmware/qmk/keyboards/electronics/splitkb36/keymaps/default/keymap.c
@@ -1,0 +1,24 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
+// Copyright 2026 hjosugi
+
+#include QMK_KEYBOARD_H
+
+enum layer_names {
+    _BASE,
+    _LOWER,
+};
+
+const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
+    [_BASE] = LAYOUT_split_3x5_3(
+        KC_Q,    KC_W,    KC_E,    KC_R,    KC_T,                      KC_Y,    KC_U,    KC_I,    KC_O,    KC_P,
+        KC_A,    KC_S,    KC_D,    KC_F,    KC_G,                      KC_H,    KC_J,    KC_K,    KC_L,    KC_SCLN,
+        KC_Z,    KC_X,    KC_C,    KC_V,    KC_B,                      KC_N,    KC_M,    KC_COMM, KC_DOT,  KC_SLSH,
+                                   KC_LCTL, KC_LALT, MO(_LOWER),       KC_SPC,  KC_ENT,  KC_BSPC
+    ),
+    [_LOWER] = LAYOUT_split_3x5_3(
+        KC_1,    KC_2,    KC_3,    KC_4,    KC_5,                      KC_6,    KC_7,    KC_8,    KC_9,    KC_0,
+        KC_ESC,  KC_TAB,  KC_MINS, KC_EQL,  KC_LBRC,                   KC_RBRC, KC_LEFT, KC_DOWN, KC_UP,   KC_RGHT,
+        KC_CAPS, KC_GRV,  KC_BSLS, KC_TRNS, KC_TRNS,                   KC_TRNS, KC_HOME, KC_PGDN, KC_PGUP, KC_END,
+                                   KC_TRNS, KC_TRNS, KC_TRNS,          KC_TRNS, KC_TRNS, KC_DEL
+    ),
+};